Integer type refactoring

+++ src/App.cpp
@@ -91,16 +91,16 @@ int App::run(int argc, char** argv)
 	TCLAP::SwitchArg interactiveSwitch("", "interactive", "View the scene interactively", cmd, false);
 	TCLAP::SwitchArg nonInteractiveSwitch("", "non-interactive", "Force non-interactive mode", cmd, false);
 	TCLAP::SwitchArg enableOpenCLSwitch("", "opencl", "Use OpenCL for raytracing", cmd, false);
-	TCLAP::ValueArg<size_t> clPlatformIdArg("", "cl-platform", "OpenCL platform to use", false, 0, "int", cmd);
+	TCLAP::ValueArg<uint64_t> clPlatformIdArg("", "cl-platform", "OpenCL platform to use", false, 0, "int", cmd);
 	TCLAP::ValueArg<uint64_t> clDeviceTypeArg("", "cl-device-type", "OpenCL device type to use", false, 0, "int", cmd);
-	TCLAP::ValueArg<size_t> clDeviceIdArg("", "cl-device", "OpenCL device to use", false, 0, "int", cmd);
+	TCLAP::ValueArg<uint64_t> clDeviceIdArg("", "cl-device", "OpenCL device to use", false, 0, "int", cmd);
 	TCLAP::ValueArg<uint64_t> testSceneArg("", "test", "Select one of the test scenes", false, 0, "int", cmd);
 	TCLAP::SwitchArg nonTestSceneSwitch("", "non-test", "Force test scene mode off", cmd, false);
 	TCLAP::SwitchArg clientSwitch("", "client", "Enable network client mode", cmd, false);
 	TCLAP::SwitchArg serverSwitch("", "server", "Enable network server mode", cmd, false);
 	TCLAP::ValueArg<std::string> sceneFileNameArg("s", "scene", "Path to the scene file", false, "", "string", cmd);
-	TCLAP::ValueArg<size_t> widthArg("w", "width", "Width of the output image or window", false, 0, "int", cmd);
-	TCLAP::ValueArg<size_t> heightArg("h", "height", "Height of the output image or window", false, 0, "int", cmd);
+	TCLAP::ValueArg<uint64_t> widthArg("w", "width", "Width of the output image or window", false, 0, "int", cmd);
+	TCLAP::ValueArg<uint64_t> heightArg("h", "height", "Height of the output image or window", false, 0, "int", cmd);
 	TCLAP::ValueArg<std::string> outputFileNameArg("o", "output", "Path to the output image file", false, "", "string", cmd);
 	TCLAP::SwitchArg autoViewSwitch("", "view", "Open the image automatically after completion", cmd, false);
